backend/application/utils.py
import openai
import faiss
import chromadb
import pinecone

openai.api_key = "openai_api_key"
pinecone.init(api_key="pinecone_api_key")

def generate_openai_response(prompt: str) -> str:
	response = openai.Completion.create(
		engine="text-davinci-003",
		prompt=prompt,
		max_tokens=100
	)
	return response.choices[0].text.strip()

def retrieve_documents(query: str):
	# Implement retrieval using FAISS, ChromaDB, or Pinecone
	# Example for FAISS:
	index = faiss.IndexFlatL2(128)
	# Add your document vectors to the index
	# Perform search and return results
	return ["Document 1", "Document 2"]
