requirements.txt
fastapi
flask
django
openai
langchain
transformers
faiss-cpu
chromadb
pinecone-client
sqlalchemy
psycopg2-binary
docker
kubernetes
boto3
azure
